// frontend/src/app/api/model/adminCreateUserDto.ts
/**
 * CoreData API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AdminCreateUserDto { 
    email?: string | null;
    password?: string | null;
    firstName?: string | null;
    lastName?: string | null;
    isActive?: boolean;
    roleId?: string | null;
}