// frontend/src/app/api/model/roleDto.ts
/**
 * CoreData API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RoleDto { 
    id?: string;
    name?: string | null;
    key?: string | null;
    rank?: number;
    isSystem?: boolean;
    permissions?: number;
}