// frontend/src/app/api/model/adminCreateUserDtoBulkItem.ts
/**
 * CoreData API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AdminCreateUserDto } from './adminCreateUserDto';


export interface AdminCreateUserDtoBulkItem { 
    key?: string | null;
    data?: AdminCreateUserDto;
}