// frontend/src/app/api/model/bulkCreateUsersRequest.ts
/**
 * CoreData API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AdminCreateUserDtoBulkItem } from './adminCreateUserDtoBulkItem';


export interface BulkCreateUsersRequest { 
    dryRun?: boolean;
    atomic?: boolean;
    continueOnError?: boolean;
    batchSize?: number;
    items?: Array<AdminCreateUserDtoBulkItem> | null;
}